# src/AppForSEII2526.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Alquiler/GetAlquilerDetail": {
      "get": {
        "tags": [
          "Alquiler"
        ],
        "operationId": "GetAlquilerDetail",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlquilerDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlquilerDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlquilerDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Alquiler/CreateAlquiler": {
      "post": {
        "tags": [
          "Alquiler"
        ],
        "operationId": "CreateAlquiler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlquilerDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AlquilerDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AlquilerDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlquilerDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlquilerDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlquilerDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlquilarItemDTO": {
        "required": [
          "cantidad",
          "materialHerramienta",
          "nombreHerramienta",
          "precio"
        ],
        "type": "object",
        "properties": {
          "nombreHerramienta": {
            "minLength": 1,
            "type": "string"
          },
          "materialHerramienta": {
            "minLength": 1,
            "type": "string"
          },
          "cantidad": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "precio": {
            "minimum": 0,
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
...
      "AlquilerDTO": {
        "required": [
          "apellidoCliente",
          "direccionEnvio",
          "fechaFin",
          "fechaInicio",
          "metodoPago",
          "nombreCliente"
        ],
        "type": "object",
        "properties": {
          "nombreCliente": {
            "minLength": 1,
            "type": "string"
          },
          "apellidoCliente": {
            "minLength": 1,
            "type": "string"
          },
          "direccionEnvio": {
            "minLength": 1,
            "type": "string"
          },
          "precioTotal": {
            "minimum": 0,
            "type": "number",
            "format": "float"
          },
          "fechaFin": {
            "type": "string",
            "format": "date-time"
          },
          "fechaInicio": {
            "type": "string",
            "format": "date-time"
          },
          "numTelefono": {
            "type": "string",
            "nullable": true
          },
          "correoElectronico": {
            "type": "string",
            "nullable": true
          },
          "metodoPago": {
            "$ref": "#/components/schemas/tiposMetodoPago"
          },
          "alquilarItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlquilarItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AlquilerDetailDTO": {
        "required": [
          "apellidoCliente",
          "direccionEnvio",
          "fechaAlquiler",
          "fechaFin",
          "fechaInicio",
          "metodoPago",
          "nombreCliente"
        ],
        "type": "object",
        "properties": {
          "nombreCliente": {
            "minLength": 1,
            "type": "string"
          },
          "apellidoCliente": {
            "minLength": 1,
            "type": "string"
          },
          "direccionEnvio": {
            "minLength": 1,
            "type": "string"
          },
          "precioTotal": {
            "minimum": 0,
            "type": "number",
            "format": "float"
          },
          "fechaFin": {
            "type": "string",
            "format": "date-time"
          },
          "fechaInicio": {
            "type": "string",
            "format": "date-time"
          },
          "numTelefono": {
            "type": "string",
            "nullable": true
          },
          "correoElectronico": {
            "type": "string",
            "nullable": true
          },
          "metodoPago": {
            "$ref": "#/components/schemas/tiposMetodoPago"
          },
          "alquilarItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlquilarItemDTO"
            },
            "nullable": true
          },
          "fechaAlquiler": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "tiposMetodoPago": {
        "enum": [
          "TarjetaCredito",
          "PayPal",
          "Efectivo"
        ],
        "type": "string"
      }
    }
  }
}